+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.0.0
 
+* Find services associated with several tags
 * Added proxy traps
 * Freeze services and factories once they are fetched
 * Added `.factory` to define factory functions